move stdlib logging handler stuff to its own module
